ha_logi_host/pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "ha-logi-host"
version = "0.1.0"
description = "Home Assistant add-on: switch Logitech mouse host via HID++ over Unifying/Bolt receiver"
requires-python = ">=3.10"
license = "GPL-3.0-only"
dependencies = [
    "pyyaml>=6.0",
    "paho-mqtt>=2.0.0,<3.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-mock>=3.10",
    "pytest-cov>=4.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--tb=short --cov --cov-fail-under=80"

[tool.coverage.run]
source = ["logi_host"]
omit = [
    "logi_host/transport.py",
    "logi_host/__main__.py",
]

[tool.coverage.report]
show_missing = true
